Split out a dump runner out of MetadataDumper
To prepare creating an alternative main class for the Cloud Extractor,
plit out the re-usable part into `DumpRunner` and call it from
`MetadataDumper`.
