+++ dumper/app/src/test/java/com/google/edwmigration/dumper/application/dumper/connector/AbstractConnectorExecutionTest.java
@@ -21,6 +21,7 @@ import static org.junit.Assert.*;
 import com.google.common.base.MoreObjects;
 import com.google.common.base.Predicates;
 import com.google.common.io.ByteSource;
+import com.google.edwmigration.dumper.application.dumper.Main;
 import com.google.edwmigration.dumper.application.dumper.MetadataDumper;
 import com.google.edwmigration.dumper.application.dumper.task.AbstractTask;
 import com.google.edwmigration.dumper.common.io.ZipArchiveEntryByteSource;
@@ -112,7 +113,7 @@ public abstract class AbstractConnectorExecutionTest extends AbstractConnectorTe
   }
 
   public void runDumper(@Nonnull String... args) throws Exception {
-    MetadataDumper dumper = new MetadataDumper();
+    Main dumper = new Main(new MetadataDumper());
     dumper.run(args);
   }
